refactoring of folders
